+++ opengever/activity/browser/listing.py
@@ -1,12 +1,12 @@
 from five import grok
-from opengever.tabbedview import GeverTableSource
 from ftw.table.interfaces import ITableSource
 from ftw.table.interfaces import ITableSourceConfig
 from opengever.activity import _
 from opengever.activity import notification_center
 from opengever.activity.browser import resolve_notification_url
 from opengever.ogds.base.actor import Actor
-from opengever.tabbedview.browser.base import BaseListingTab
+from opengever.tabbedview import BaseListingTab
+from opengever.tabbedview import GeverTableSource
 from plone import api
 from zope.interface import implements
 from zope.interface import Interface
