+++ opengever/tabbedview/browser/tasklisting.py
@@ -6,8 +6,8 @@ from opengever.globalindex.model.task import Task
 from opengever.globalindex.utils import indexed_task_link_helper
 from opengever.ogds.base.utils import get_current_admin_unit
 from opengever.tabbedview import _
+from opengever.tabbedview import GeverTabMixin
 from opengever.tabbedview import SqlTableSource
-from opengever.tabbedview.browser.base import OpengeverTab
 from opengever.tabbedview.browser.listing import ListingView
 from opengever.tabbedview.filters import Filter
 from opengever.tabbedview.filters import FilterList
@@ -31,8 +31,7 @@ class IGlobalTaskTableSourceConfig(ITableSourceConfig):
     """
 
 
-class GlobalTaskListingTab(grok.View, OpengeverTab,
-                           ListingView):
+class GlobalTaskListingTab(grok.View, GeverTabMixin, ListingView):
     """A tabbed view mixin which brings support for listing tasks from
     the SQL (globally over all clients).
