+++ opengever/tabbedview/browser/users.py
@@ -4,8 +4,8 @@ from opengever.base.model import create_session
 from opengever.ogds.base.actor import Actor
 from opengever.ogds.models.user import User
 from opengever.tabbedview import _
+from opengever.tabbedview import GeverTabMixin
 from opengever.tabbedview import SqlTableSource
-from opengever.tabbedview.browser.base import OpengeverTab
 from opengever.tabbedview.browser.listing import ListingView
 from opengever.tabbedview.helper import boolean_helper
 from opengever.tabbedview.helper import email_helper
@@ -30,7 +30,7 @@ class IUsersListingTableSourceConfig(ITableSourceConfig):
     """
 
 
-class UsersListing(grok.View, OpengeverTab, ListingView):
+class UsersListing(grok.View, GeverTabMixin, ListingView):
     """Tab registered on contacts folder (see opengever.contact) listing all
     users.
     """
