+++ opengever/task/browser/overview.py
@@ -2,14 +2,14 @@ from Acquisition import aq_inner
 from Acquisition import aq_parent
 from five import grok
 from opengever.base.browser.helper import get_css_class
-from opengever.tabbedview.browser.base import OpengeverTab
+from opengever.tabbedview import GeverTabMixin
 from opengever.task import _
 from opengever.task.task import ITask
 from plone import api
 from Products.CMFCore.utils import getToolByName
 
 
-class Overview(grok.View, OpengeverTab):
+class Overview(grok.View, GeverTabMixin):
     grok.context(ITask)
     grok.name('tabbedview_view-overview')
     grok.template('overview')
